Abstract `fetchAndUpdateDCData`
In order to call it for different attributes (block and background)
